Update
Added an annotation to the Pie chart that tell you what data the Pie chart is using.
Changed the text on the buttons to be more concise.
Changed the functionality of the buttons so that the button that corresponds to the current dataset is disabled.
Added a stylesheet.
Added padding to the Pie chart to distance it from the buttons.